requirements.txt
python-slugify==4.0.1
hdx-python-api==4.8.3
-r docker-requirements.txt